refactor(cli): split share oclif commands
